==> onehot_encoding/3.augmentation.ipynb <==
import numpy as np # linear algebra
import pandas as pd # data processing, CSV file I/O (e.g. pd.read_csv)
import os
import joblib
import sklearn
# Import Scikit-Learn library for decision tree models
import sklearn 

from sklearn import datasets, linear_model, metrics, tree
from sklearn.utils import resample

from sklearn.tree import DecisionTreeRegressor

from sklearn.ensemble import BaggingRegressor, GradientBoostingRegressor


from sklearn.neighbors import KNeighborsRegressor

        
# Dataset splitting        
from sklearn.model_selection import train_test_split


from matplotlib import pyplot as plt


import os
import joblib

df = pd.read_csv("dataset_scaled.csv")
df

input_col = [col for col in df.columns]
input_col.remove('Unnamed: 0')
input_col.remove('size_nm')
input_col.remove('g_factor*10^4.1')
input_col

# Load ML model for predicting size

loaded_rf_size = joblib.load('model_aug_size_DecisionTree.joblib')

# Replace 'None' entries in 'size_nm' column by predicted values.

a = 0
for index, row in df.iterrows():
    if row['size_nm'] == 'None':
        X = df.loc[index, input_col].to_numpy()
        df.loc[index, 'size_nm'] = loaded_rf_size.predict(X.reshape(1, -1))[0]
        a += 1

        
# Save the dataset where all 'None' values in 'Absorbane max (nm)' column are replaced.   
df.to_csv('dataset_augmented_size.csv')
df_first_aug = pd.read_csv('dataset_augmented_size.csv')

df_first_aug

#Saves the row indexes to drop in the column of g_factor into a list
total_row_num = len(df_first_aug)
drop_list_g =[]
for row_i in range(total_row_num):
    if df_first_aug['g_factor*10^4.1'].values[row_i] == 'None':
        drop_list_g.append(row_i)
    
len(drop_list_g)

#Drops the appropriate rows
df_g_scaled_encoded = df_first_aug.drop(drop_list_g)
#Saves the data for absorbance modelling to CSV
df_g_scaled_encoded.to_csv('dataset_scaled_g.csv')

## Move to 'g_aug_model' notebook to optimize model for augmenting Absorbance

# adding column 'size_nm' to input
input_col.append('size_nm')
input_col_2 = input_col.copy()
input_col.remove('size_nm')
print(input_col_2)

# Now, the dataset only has 'None' values in the 'Abs_nm' column

df_g = pd.read_csv('dataset_augmented_size.csv')

# Load ML model for predicting Aborbance

loaded_rf_g = joblib.load('model_aug_g_DecisionTree.joblib')


# # Replace 'None' entries in 'Abs_nm' column by predicted values.

a = 0
for index, row in df_g.iterrows():
    if row['g_factor*10^4.1'] == 'None':
        X = df_g.loc[index, input_col_2].to_numpy()
        df_g.loc[index, 'g_factor*10^4.1'] = loaded_rf_g.predict(X.reshape(1, -1))[0]
        a += 1

# # # Save the dataset where all 'None' values are replaced.

# # # Ready to use for other analysis.

df_g.to_csv('dataset_augmented_all.csv')
